requirements.txt
selenium==4.15.0
requests==2.31.0
pytest==7.4.0
allure-pytest==2.13.2
pytest-xdist==3.3.1
webdriver-manager==4.0.0
python-dotenv==1.0.0